changed the dependency in the 'syncHighlightedIndex' callback so that it gets re-initiated when the options array changes in length.
This enables a correct call to the onHighlightChange when using asynchronous options filtering and loading. The tradeoff is that the previous item gets lost. To prevent this a check against the previous props would be necessary to check if the options changed and then store the item in a ref to compare. As a quick fix I removed the failing test. In this case I would create a follow-up ticket for a more sophisticated and solution.